refactor(Template): 상수를 템플릿화한 파일의 상위 디렉토리 명을 Template으로 변경
